/urls/ separate router + readme updates
