+++ app/templates/index.html
@@ -11,7 +11,7 @@
   <main>
     <div class="container mx-auto px-6">
       <c-toast></c-toast>
-      <c-form origin="{{ origin }}" url="/api/v1/shorten"></c-form>
+      <c-form origin="{{ origin }}" url="/api/v1/urls/shorten"></c-form>
     </div>
   </main>
 </body>
