+++ frontend/index.html
@@ -9,7 +9,7 @@
   <main>
     <div class="container mx-auto px-6">
       <c-toast></c-toast>
-      <c-form origin="http://localhost:8000" url="/api/v1/shorten"></c-form>
+      <c-form origin="http://localhost:8000" url="/api/v1/urls/shorten"></c-form>
     </div>
   </main>
   <script type="module" src="/src/main.ts"></script>
